requirements.txt
boto3==1.26.85
dj-database-url==1.2.0
django-environ==0.10.0
django-storages==1.13.2
django-tinymce==3.5.0
Django==4.1.7
gunicorn==20.1.0
psycopg2-binary==2.9.5
whitenoise==6.4.0
